refactor: Move rule pattern validation to utility functions
- Removed the validateRulePattern method from RulesService and integrated it into RulesUtils.
- Enhanced validation to support URL patterns alongside domain and IP address formats.
- Updated the validation logic for host and path patterns to improve clarity and maintainability.
