+++ extension/src/services/RulesService.js
@@ -5,7 +5,7 @@
 
 import { storageManager } from "../config/StorageManager.js"
 import { tabGroupState } from "../state/TabGroupState.js"
-import { isIPv4Pattern, matchIPv4 } from "../utils/RulesUtils.js"
+import { isIPv4Pattern, matchIPv4, validateRulePattern } from "../utils/RulesUtils.js"
 
 class RulesService {
   /**
@@ -359,7 +359,7 @@ class RulesService {
           break
         }
 
-        const validation = this.validateRulePattern(pattern.trim())
+        const validation = validateRulePattern(pattern.trim())
         if (!validation.isValid) {
           errors.push(`Invalid pattern "${pattern}": ${validation.error}`)
         }
@@ -370,270 +370,6 @@ class RulesService {
       isValid: errors.length === 0,
       errors
     }
-  }
-
-  /**
-   * Validates a rule pattern (domain, IP, or URL)
-   * @param {string} pattern - Pattern to validate
-   * @returns {Object} Validation result with isValid and error
-   */
-  validateRulePattern(pattern) {
-    if (!pattern || typeof pattern !== "string") {
-      return { isValid: false, error: "Pattern must be a string" }
-    }
-
-    const cleanPattern = pattern.toLowerCase().trim()
-
-    if (cleanPattern.length === 0) {
-      return { isValid: false, error: "Pattern cannot be empty" }
-    }
-
-    if (cleanPattern.length > 300) {
-      return { isValid: false, error: "Pattern too long (max 300 characters)" }
-    }
-
-    // Check if it's a URL pattern or host pattern
-    const hasPath = cleanPattern.includes("/")
-    const [hostPattern, pathPattern] = hasPath ? cleanPattern.split("/", 2) : [cleanPattern, ""]
-
-    // Validate host part (domain or IP)
-    const hostValidation = this.validateHostPattern(hostPattern)
-    if (!hostValidation.isValid) {
-      return hostValidation
-    }
-
-    // Validate path part if present
-    if (hasPath) {
-      const pathValidation = this.validatePathPattern(pathPattern)
-      if (!pathValidation.isValid) {
-        return pathValidation
-      }
-    }
-
-    return { isValid: true }
-  }
-
-  /**
-   * Validates a host pattern (domain or IP)
-   * @param {string} pattern - Host pattern to validate
-   * @returns {Object} Validation result with isValid and error
-   */
-  validateHostPattern(pattern) {
-    if (!pattern) {
-      return { isValid: false, error: "Host pattern cannot be empty" }
-    }
-
-    // Check if it's an IPv4 pattern
-    if (isIPv4Pattern(pattern)) {
-      return this.validateIPv4Pattern(pattern)
-    }
-
-    // Fall back to domain validation
-    return this.validateDomainPattern(pattern)
-  }
-
-  /**
-   * Validates an IPv4 pattern
-   * @param {string} pattern - IPv4 pattern to validate
-   * @returns {Object} Validation result with isValid and error
-   */
-  validateIPv4Pattern(pattern) {
-    if (!pattern) {
-      return { isValid: false, error: "IPv4 pattern cannot be empty" }
-    }
-
-    const parts = pattern.split(".")
-
-    if (parts.length !== 4) {
-      return { isValid: false, error: "IPv4 address must have exactly 4 octets" }
-    }
-
-    for (let i = 0; i < parts.length; i++) {
-      const part = parts[i]
-
-      if (part === "*") {
-        continue // Wildcard is valid
-      }
-
-      if (!/^\d+$/.test(part)) {
-        return { isValid: false, error: `Invalid IPv4 octet "${part}". Must be 0-255 or *` }
-      }
-
-      const num = parseInt(part, 10)
-
-      if (num < 0 || num > 255) {
-        return { isValid: false, error: `IPv4 octet "${part}" must be between 0 and 255` }
-      }
-
-      if (part !== num.toString()) {
-        return { isValid: false, error: `IPv4 octet "${part}" cannot have leading zeros` }
-      }
-    }
-
-    return { isValid: true }
-  }
-
-  /**
-   * Validates a domain pattern
-   * @param {string} pattern - Domain pattern to validate
-   * @returns {Object} Validation result with isValid and error
-   */
-  validateDomainPattern(pattern) {
-    if (!pattern) {
-      return { isValid: false, error: "Domain pattern cannot be empty" }
-    }
-
-    // Check for ** wildcard (TLD wildcard)
-    if (pattern.includes("**")) {
-      const parts = pattern.split("**")
-      if (parts.length !== 2) {
-        return { isValid: false, error: "Invalid ** pattern. Use format: domain.** or *.domain.**" }
-      }
-
-      const prefix = parts[0]
-      const suffix = parts[1]
-
-      // Validate prefix
-      if (!prefix || !prefix.endsWith(".")) {
-        return {
-          isValid: false,
-          error: "** pattern must have domain prefix ending with dot (e.g., google.)"
-        }
-      }
-
-      // Check if prefix has subdomain wildcard
-      if (prefix.startsWith("*.")) {
-        const basePrefix = prefix.substring(2)
-        if (!basePrefix || !basePrefix.match(/^[a-zA-Z0-9.-]+\.$/)) {
-          return { isValid: false, error: "Invalid subdomain wildcard with ** pattern" }
-        }
-      } else {
-        // Validate prefix without subdomain wildcard
-        if (!prefix.match(/^[a-zA-Z0-9.-]+\.$/)) {
-          return { isValid: false, error: "Invalid domain prefix in ** pattern" }
-        }
-      }
-
-      // Validate suffix (usually empty, but could be something like .co in future)
-      if (suffix && !suffix.match(/^[a-zA-Z0-9.-]*$/)) {
-        return { isValid: false, error: "Invalid suffix in ** pattern" }
-      }
-
-      return { isValid: true }
-    }
-
-    // Check for * wildcard (subdomain wildcard)
-    if (pattern.startsWith("*.")) {
-      const baseDomain = pattern.substring(2)
-
-      if (!baseDomain || !baseDomain.includes(".")) {
-        return { isValid: false, error: "Invalid * pattern. Use format: *.domain.com" }
-      }
-
-      if (baseDomain.includes("*")) {
-        return { isValid: false, error: "Multiple wildcards not allowed in domain pattern" }
-      }
-
-      if (!/^[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/.test(baseDomain)) {
-        return { isValid: false, error: "Invalid base domain in * pattern" }
-      }
-
-      return { isValid: true }
-    }
-
-    // Regular domain validation
-    if (!/^[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/.test(pattern)) {
-      return { isValid: false, error: "Invalid domain format" }
-    }
-
-    // Check for invalid patterns
-    if (pattern.startsWith(".") || pattern.endsWith(".")) {
-      return { isValid: false, error: "Domain cannot start or end with a dot" }
-    }
-
-    if (pattern.includes("..")) {
-      return { isValid: false, error: "Domain cannot contain consecutive dots" }
-    }
-
-    if (pattern.startsWith("-") || pattern.endsWith("-")) {
-      return { isValid: false, error: "Domain cannot start or end with a hyphen" }
-    }
-
-    return { isValid: true }
-  }
-
-  /**
-   * Validates a path pattern
-   * @param {string} pattern - Path pattern to validate
-   * @returns {Object} Validation result with isValid and error
-   */
-  validatePathPattern(pattern) {
-    if (!pattern) {
-      return { isValid: false, error: "Path pattern cannot be empty" }
-    }
-
-    // Remove leading slash if present
-    const cleanPattern = pattern.startsWith("/") ? pattern.substring(1) : pattern
-
-    if (cleanPattern.length === 0) {
-      return { isValid: false, error: "Path pattern cannot be empty" }
-    }
-
-    if (cleanPattern.length > 100) {
-      return { isValid: false, error: "Path pattern too long (max 100 characters)" }
-    }
-
-    // Check for ** wildcard in path
-    if (cleanPattern.includes("**")) {
-      const parts = cleanPattern.split("**")
-      if (parts.length !== 2) {
-        return { isValid: false, error: "Invalid ** pattern in path. Use format: prefix/**/suffix" }
-      }
-
-      const prefix = parts[0]
-      const suffix = parts[1]
-
-      // Validate prefix (if present)
-      if (prefix && !this.isValidPathSegment(prefix)) {
-        return { isValid: false, error: "Invalid prefix in path ** pattern" }
-      }
-
-      // Validate suffix (if present)
-      if (suffix && !this.isValidPathSegment(suffix)) {
-        return { isValid: false, error: "Invalid suffix in path ** pattern" }
-      }
-
-      return { isValid: true }
-    }
-
-    // Basic path validation - allow alphanumeric, hyphens, underscores, dots, slashes, and asterisks
-    if (!/^[a-zA-Z0-9._/*-]+$/.test(cleanPattern)) {
-      return { isValid: false, error: "Path pattern contains invalid characters" }
-    }
-
-    // Check for invalid patterns
-    if (cleanPattern.includes("//")) {
-      return { isValid: false, error: "Path pattern cannot contain consecutive slashes" }
-    }
-
-    return { isValid: true }
-  }
-
-  /**
-   * Validates a path segment (helper for path validation)
-   * @param {string} segment - Path segment to validate
-   * @returns {boolean} True if segment is valid
-   */
-  isValidPathSegment(segment) {
-    if (!segment) return true // Empty segments are allowed
-
-    // Remove leading/trailing slashes
-    const cleanSegment = segment.replace(/^\/+|\/+$/g, "")
-
-    if (cleanSegment.length === 0) return true
-
-    // Allow alphanumeric, hyphens, underscores, dots, and slashes
-    return /^[a-zA-Z0-9._/-]+$/.test(cleanSegment) && !cleanSegment.includes("//")
   }
 
   /**
